Some refactoring.

+++ GraphCount/GraphCount/Algorithm.h
@@ -3,8 +3,8 @@
 #ifndef GRAPH_ALGORITHM_H
 #define GRAPH_ALGORITHM_H
 
-int mActually(Graph graph);
-int oneDirectedNum(Graph graph);
+long mActually(Graph graph);
+long oneDirectedNum(Graph graph);
 
 #endif //GRAPH_ALGORITHM_H
